Don't include the pid in the dbus path when on flatpak
Summary:
Flatpak requests us to list the services that will be exposed to the outside,
better have it the name of the application without a random number in the end.
In these systems the pid doesn't add much anyway, because two instances of
the application will always end up with the same pid anyway, from the point
of view of the application.

Test Plan: Applied the patch to `flatpak-kde-runtime`, fixes the issue for katomic and blinken.

Reviewers: #frameworks, jgrulich, aacid

Reviewed By: aacid

Subscribers: dfaure, davidedmundson, aacid

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D5775