Introduce UpdateLaunchEnvJob
This is a class that exists currently in plasma-workspace/libkworkspace
and has proved useful.

Effectively it abstracts setting an environment variable into the system
for launching other applications. For example a kcminit module setting
the XCURSOR_THEME.

We now also want to use this within kwin, which means either duplicating
the code or moving to a library.

In moving we dropped subclassing KJob because of reasons because of tier
issues. We just need a simple finished signal anyway.