Add kdedbus6, a replacement for qdbus
We use the dbus commandline tool in various places, for example for global shortcuts.

However depending on it has several problems

1) It lives in the qttools repo, together with various developer-only tools like qdbusviewer and qtdesigner
By depending on qdbus from qttools we risk pulling in a couple of develper-only tools into user installations

2) The upstream name of the tool is qdbus, however for coinstallability with Qt5 some distros rename it to e.g.
qdbus6 or qdbus-qt6. The naming is not consistent across distros, so it needs buildsystem-level intervention to find the
right name, which complicates usage a lot

See https://invent.kde.org/plasma/plasma-workspace/-/issues/60

This changes takes the code of upstream qdbus (with minor modifications to avoid using internal Qt headers) and turns it
into the kdedbus6 tool. This allows us to have consistend and predictable naming and packaging of the tool.

The tool is not called kdbus to avoid confusion with the kdbus project.