Extend timeout for --replace option
When using --replace 7431ea1f8a5f08d412e447a3ddd3d668a7c1ca28 changed
things so that we don't just wait for the app to start quitting but to
finish quitting.

This could take longer, especially if the other app is crashing on exit
and the timeout we need should be extended appropriately.

There's no harm in a long timeout as:
 1) we still proceed immediately when things are ready
 2) The --replace feature exists as a dev tool